+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ### Unreleased
 
+* Fixed a code generation compilation error where a component contains a `@PostDispose` annotated method, no `@PreDispose` annotated method and is annotated with `@ArezComponent( disposeNotifier = Feature.DISABLE )`.
+
 ### [v0.142](https://github.com/arez/arez/tree/v0.142) (2019-07-15)
 [Full Changelog](https://github.com/arez/arez/compare/v0.141...v0.142)
 [API Differences](https://arez.github.io/api-diff/?key=arez&old=0.141&new=0.142)